+++ src/gui/Splash.qml
@@ -133,7 +133,7 @@ Window {
                         font.letterSpacing: 4
                     }
                     Text {
-                        text: "V E R S I O N  4 . 2 2 . 5"
+                        text: "V E R S I O N  4 . 2 2 . 6"
                         color: "#666"
                         font.pixelSize: 10
                         font.bold: true